test: :white_check_mark: move helper module
